Fix test build break in netfx
